{"schema":"decodex/commit/1","summary":"Split diversity selection modules","authority":"manual"}
